fix(domain):修改id属性为Integer

+++ asset/src/main/java/www/topview/asset/domain/po/Asset.java
@@ -1,5 +1,6 @@
 package www.topview.asset.domain.po;
 
+import com.baomidou.mybatisplus.annotation.IdType;
 import com.baomidou.mybatisplus.annotation.TableId;
 import com.baomidou.mybatisplus.annotation.TableName;
 import lombok.Data;
@@ -19,7 +20,7 @@ public class Asset {
      * 主键
      */
     @TableId
-    private Long id;
+    private Integer id;
 
     /**
      * 资产名字
